fix(blur): when switching video tracks.
